requirements.txt
streamlit>=1.28.0
opencv-python-headless==4.5.3.56
pytesseract>=0.3.10
pillow>=9.5.0
numpy==1.23.5
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0